requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
schedule>=1.2.0
python-dotenv>=1.0.0
openai>=1.0.0
newspaper3k>=0.2.8